Move gcs over to download service.
